+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -278,6 +278,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fix diffusion profile displayed in the inspector.
 - Fixed override camera rendering custom pass API aspect ratio issue when rendering to a render texture.
 - Fixed the incorrect value written to the VT feedback buffer when VT is not used.
+- Fixed support for ray binning for ray tracing in XR (case 1346374).
 
 ### Changed
 - Changed Window/Render Pipeline/HD Render Pipeline Wizard to Window/Rendering/HDRP Wizard